Fix file picker error on Linux (#97)
* Replace QGuiApplication with QApplication

Replaces the usage of a QGuiApplication with a QApplication. This allows
for Qt widgets to be used, such as the FileDialog widget.

* Replace instances of "file:///" with "file://"

This fixes an issue on Unix-based systems where the leading slash in
absolute path names was being removed, causing things like ROM
libraries to be stored as relative paths.

* Add helper functions for file string prefixes

Adds helper functions to help deal with the different number of slashes
in "file:///" prefixes.

* Add missing parameter

* Fix Linux build error

* Convert new QGuiApplication objects to QApplications

* Use helper functions for all file URI handling